requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==1.11.16
django-filter==2.0.0
djangorestframework==3.9.0
httpie==1.0.2
idna==2.7
Markdown==3.0.1
Pygments==2.3.0
pytz==2018.7
requests==2.20.1
urllib3==1.24.1
